rename status badge
