+++ packages/react/src/status-badge/docs/code.mdx
@@ -1,0 +1,3 @@
+## Props
+
+<ComponentPropsTable name="StatusBadge" />
