+++ packages/react/src/status-badge/index.ts
@@ -1,0 +1,1 @@
+export * from './StatusBadge';
